+++ examples/typeScript/reports/getReport.ts
@@ -6,21 +6,26 @@ import * as MwsApi from 'amazon-mws';
 const amazonMws = new MwsApi();
 amazonMws.setApiKey(accessKey, accessSecret);
 
+/**
+ * This will not provide you Throttling details in Header.
+ * Amazon MWS itself not providing Throttling detail in GetReport call.
+ */
+
 const reportRequest = async () => {
 
-    let response = await amazonMws.reports.search({
-        'Version': '2009-01-01',
-        'Action': 'GetReport',
-        'SellerId': 'SELLER_ID',
-        'MWSAuthToken': 'MWS_AUTH_TOKEN',
-        'ReportId': 'REPORT_ID'
-    }).catch(error => {
-        if (error) {
-            console.log('error ', error);
-            return;
-        }
-    });
-    console.log('response', response);
+    try {
+        const response: any = await amazonMws.reports.search({
+            'Version': '2009-01-01',
+            'Action': 'GetReport',
+            'SellerId': 'SELLER_ID',
+            'MWSAuthToken': 'MWS_AUTH_TOKEN',
+            'ReportId': 'REPORT_ID'
+        });
+        console.log('response', response);
+    } catch (error: any) {
+        console.log('error ', error);
+    }
+
 };
 
 reportRequest();